+++ changelog.md
@@ -6,9 +6,11 @@ dev
 ===
 
 - Merge ProceduralRule class with Rule
-- Ignore was/were differences in Predicate content
+- Ignore was/were differences in Predicate content text
 - Let input JSON label a Rule as the "exclusive" way to get output
 - Create addition operator for Factors and Rules
+- Let Rule init method handle the necessary Procedure init method
+- Use addition operator to add Factors as Rule inputs
 
 0.1.0 (2019-06-10)
 ==================